TCL and non-TCL tests can now be run in the same queue.

+++ src/GUI/Application/QueueRunner.py
@@ -44,31 +44,37 @@ class QueueRunner(Thread):
         # TODO catch any exceptions run by prober and try to continue, but only if a flag in the __init__ has been
         # set
         self.queue.schedule_experiments()
-        for i in range(len(self.queue)):
+        i = 0
+        while i < len(self.queue):
             self.current_experiment = self.queue.get_ith_experiment(i)
-            self.experiment_status[self.current_experiment] = 0
-            # The file to store the json to call prober on to run the experiment
+            print(self.current_experiment.get_name())
+            if self.current_experiment.get_name()[-5:] == '(Tcl)':
+                tcl_end = i + 1
+                while tcl_end < len(self.queue) and self.queue.get_ith_experiment(tcl_end).get_name()[-5:] == '(Tcl)':
+                    tcl_end += 1
+                self.run_tcl_tests(tcl_end, i)
+                i = tcl_end
+            else:
+                self.experiment_status[self.current_experiment] = 0
+                # The file to store the json to call prober on to run the experiment
 
-            tmp_file_name = self.tmp_dir + "/tmp" + self.current_experiment.get_name().replace(" ", "_") + ".json"
-            self.current_experiment.export_to_json(tmp_file_name)
-            Prober.main(["Prober.py", tmp_file_name], config_manager=Globals.systemConfigManager, queue_result=self.queue_result)
-            self.experiment_status[self.current_experiment] = 1
-            os.remove(tmp_file_name)
+                tmp_file_name = self.tmp_dir + "/tmp" + self.current_experiment.get_name().replace(" ", "_") + ".json"
+                self.current_experiment.export_to_json(tmp_file_name)
+                Prober.main(["Prober.py", tmp_file_name], config_manager=Globals.systemConfigManager, queue_result=self.queue_result)
+                self.experiment_status[self.current_experiment] = 1
+                os.remove(tmp_file_name)
+                i += 1
         self.current_experiment = None
         self.queue_result.end_queue()
         self.queue_result.save()
         print "\n====================\nQueue has finished\n====================\n"
 
-    def run_tcl_tests(self, target_location="C:\\Users\\ofs9424\\SPAE\\TCL\\combined.tcl"):
+    def run_tcl_tests(self, tcl_end, start_index=0, target_location="C:\\Users\\ofs9424\\SPAE\\TCL\\combined.tcl"):
         """
-                Run the provided queue
+                Run a series of TCL tests starting at the given index
                 :return:
                     Nothing
                 """
-        print "\n====================\nStarting the Queue\n====================\n"
-        self.queue_result.start_queue()
-        # TODO catch any exceptions run by prober and try to continue, but only if a flag in the __init__ has been
-        # set
         self.queue.schedule_experiments()
         now = datetime.datetime.today()
 
@@ -77,27 +83,28 @@ class QueueRunner(Thread):
         self.queue_result.time = now
         result_dir = Globals.systemConfigManager.get_results_manager().results_root
         master_tcl = ""
-        for i in range(len(self.queue)):
+        for i in range(start_index, tcl_end):
             with open(self.queue.get_ith_experiment(i).tcl_file, "r") as f:
                 done_sets = False
                 master_tcl += "set output \"" + result_dir + "/" + name + "/Collected_Data" + str(i+1) + ".csv\"\n"
                 master_tcl += "set index " + str(i+1) + "\n"
                 for line in f.readlines():
                     if not done_sets:
-                        if not line.startswith("set"):
-                            done_sets = True
-                        else:
-                            words = line.split(' ')
-                            val = words[1]
-                            for k in self.queue.get_ith_experiment(i).config_dict['Data'].keys():
-                                if k == val:
-                                    line = words[0]+" "+words[1]+" "+str(self.queue.get_ith_experiment(i).config_dict['Data'][k]) + "\n"
+                        if not line.startswith("#"):
+                            if not line.startswith("set"):
+                                done_sets = True
+                            else:
+                                words = line.split(' ')
+                                val = words[1]
+                                for k in self.queue.get_ith_experiment(i).config_dict['Data'].keys():
+                                    if k == val:
+                                        line = words[0]+" "+words[1]+" "+str(self.queue.get_ith_experiment(i).config_dict['Data'][k]) + "\n"
                     master_tcl += line
         with open(target_location, "w") as f:
             f.writelines(master_tcl)
 
         master_experiment = Experiment("..\\..\\template.json")
-        for i in range(len(self.queue)):
+        for i in range(start_index, tcl_end):
             if self.queue.get_ith_experiment(i).config_dict.get('Devices', None) is not None:
                 for key in self.queue.get_ith_experiment(i).config_dict['Devices']:
                     if key not in master_experiment.config_dict['Devices']:
@@ -116,26 +123,9 @@ class QueueRunner(Thread):
         Prober.main(["Prober.py", tmp_file_name], config_manager=Globals.systemConfigManager,
                     queue_result=self.queue_result)
 
-        # for i in range(len(self.queue)):
-        #     self.current_experiment = self.queue.get_ith_experiment(i)
-        #     self.experiment_status[self.current_experiment] = 0
-        #     # The file to store the json to call prober on to run the experiment
-        #     if self.current_experiment.config_dict.get('Experiment', None) is not None:
-        #         self.current_experiment.config_dict['Data']['Tcl_Script_Location'] = target_location
-        #         tmp_file_name = self.tmp_dir + "/tmp" + self.current_experiment.get_name().replace(" ", "_") + ".json"
-        #         self.current_experiment.export_to_json(tmp_file_name)
-        #         Prober.main(["Prober.py", tmp_file_name], config_manager=Globals.systemConfigManager, queue_result=self.queue_result)
-        #         self.experiment_status[self.current_experiment] = 1
-        #         os.remove(tmp_file_name)
-        # self.current_experiment = None
-
-        self.queue_result.end_queue()
-        self.queue_result.save()
-        print "\n====================\nQueue has finished\n====================\n"
-
     def run(self):
-        self.run_tcl_tests()
-        #self.run_standard()
+        #self.run_tcl_tests()
+        self.run_standard()
 
     def get_current_experiment(self):
         """
